Add `HalfEdgeBuilder::update_from_other_edge`
